Move migration to its own file
